% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgeowheels.R
\name{install_wheel}
\alias{install_wheel}
\title{Install Python Wheels From 'geospatial-wheels' Repository}
\usage{
install_wheel(
  package,
  version = "latest",
  pyversion = "latest",
  architecture = "win_amd64",
  python = get_rgeowheels_python(),
  destdir = tempdir(),
  url_only = FALSE,
  download_only = FALSE
)
}
\arguments{
\item{package}{Python package name to install. e.g. \code{"rasterio"}}

\item{version}{Python package version to install. Default \code{"latest"} determines latest version available from asset list (considers \code{pyversion} if set).}

\item{pyversion}{Python version to install package for. Default \code{"latest"} determines latest version available from asset list.}

\item{architecture}{Python package version to install. Default \code{"win_amd64"}, alternatives include \verb{"win_arm64}" and \code{"win32"}.}

\item{python}{Path to Python executable to use for install. Default: \code{get_rgeowheels_python()}}

\item{destdir}{Destination directory for downloaded wheel file. Default: \code{tempdir()}}

\item{url_only}{Return the URL of the .whl file without downloading? Default: \code{FALSE}}

\item{download_only}{Download .whl file without attempting install? Default: \code{FALSE}}
}
\value{
Called for side effects (download and install a Python wheel). Returns \emph{character} containing path to .whl file when \code{url_only=TRUE} or \code{download_only=TRUE}.
}
\description{
Used to download and install the latest versions of wheels available from \url{https://github.com/cgohlke/geospatial-wheels}.
}
