% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgeowheels.R
\name{list_assets}
\alias{list_assets}
\title{List assets available from "geospatial-wheels" repository}
\usage{
list_assets(release = NULL, update_cache = FALSE)
}
\arguments{
\item{release}{Specify custom release to list assets for. Default: \code{NULL}}

\item{update_cache}{Force update of wheel download index? Default: \code{FALSE}}
}
\value{
A \emph{data.frame} containing  \code{package}, \code{version}, \code{pyversion}, \code{architecture} and other metadata about each asset in a release.
}
\description{
List assets available from "geospatial-wheels" repository
}
