% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgeowheels.R
\name{refresh_rgeowheels_cache}
\alias{refresh_rgeowheels_cache}
\title{Refresh rgeowheels cache}
\usage{
refresh_rgeowheels_cache()
}
\value{
Called for side effects. Updates the local cache with the latest available wheels.
}
\description{
Force update of the cached wheel download index from the latest GitHub release.
}
