% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgeowheels-package.R
\docType{package}
\name{rgeowheels-package}
\alias{rgeowheels}
\alias{rgeowheels-package}
\title{rgeowheels: Convenient Access to 'Python' 'Wheel' Packages Prepared for 'Windows'}
\description{
Downloads pre-compiled 'Windows' 'wheel' files (.whl) for 'Python' geospatial packages from \url{https://github.com/cgohlke/geospatial-wheels}. These are unofficial binary installers for some geospatial libraries prepared by Christoph Gohlke. Wheels for various packages, 'Python' versions, and architectures are made available via 'GitHub' releases, providing the ability to revert to prior versions when needed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/brownag/rgeowheels}
  \item \url{https://humus.rocks/rgeowheels}
  \item Report bugs at \url{https://github.com/brownag/rgeowheels/issues}
}

}
\keyword{internal}
