% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgeowheels.R
\name{set_rgeowheels_python}
\alias{set_rgeowheels_python}
\alias{get_rgeowheels_python}
\title{Get or Set Python Path}
\usage{
set_rgeowheels_python(x)

get_rgeowheels_python()
}
\arguments{
\item{x}{Path to \code{python} or \code{python3} binary.}
}
\value{
\emph{character} Value of option \code{"rgeowheels.python"}, or, if set, the value of the system environment variable \code{"R_RGEOWHEELS_PYTHON"}. If neither are set, then the result of \code{Sys.which("python")} (or \code{Sys.which("python3")} if the former fails).
}
\description{
Set the path the Python binary used to run installation commands. May be a system or virtual/conda environment.
}
